Model Runs Version
After changing some of the dictionary calls the model runs by going to main.py and running the lines in waterpy() one by one. Also, had to create a dummy version of a csv called pptProp_24hrs.csv. I have no idea where this file is used or if it even is used for anything at this time. The post process code opens a lot of figures but doesn't resolve anything, though output is generated.