Working version for DRB
This set of code works for DRB basins and variables. Note that the xml files where the variables are stored are not on git hub, and neither is the key to convert from xml vars to waterpy vars. There are also 4 base variables files that are needed. Lastly, there is a modelconfig_template.ini file that is needed for batching runs. The code is not optimized for multi-processing, or running more than one basin.