modify geoprocessing for DRB
Geoprocessing characteristics and parameters are solved for ones that are able to be calculated. Still missing a few that need to be hard coded in. Still need to calculate twi at this point as well. Still need to modiffy the code to run on multiple processors. Updated git ignore for some of the rasters